refactor: Moved BaseEmail and its utilities to notification.email package
